Shipt coding challenge for internship consideration. Build a tic-tac-toe mobile app!
Overview

For this challenge, you will be building a tic tac toe game application. Please feel free to use any mobile framework or technology you wish. If you do not know how to build a mobile application (or do not think you have time to do it) please build a web application that will run scaled down on a mobile device.

The Game Should...

Allow 2 players to play tic tac toe (Follow the rules of tic tac toe)
Have 3x3 grid on which the players can play
Allow the players to take turns marking spaces on the 3x3 grid
Recognize when a player has won and declare that player as victorious
Allow the user to start a new game